Feature: added Internationalization support to backend Laravel

+++ app/Http/Controllers/Api/V1/Admin/SshKeyController.php
@@ -54,11 +54,28 @@ class SshKeyController extends Controller
 
     public function destroy(SshKey $sshKey)
     {
+        if ($this->sshKeyAssociated($sshKey)) {
+            return response()->json([
+                'message' => __('messages.ssh_keys.associated_servers_msg'),
+            ], 422);
+        }
+        
         $notifiable = \App\Models\User::find($sshKey->created_by);
         $sshKey->delete();
 
         Notification::send($notifiable, new ResourceDeletedNotification($sshKey, 'ssh key', 'ssh-keys'));
 
         return response()->noContent();
+    }
+
+    private function sshKeyAssociated(SshKey $sshKey)
+    {
+        $associatedServers = $sshKey->servers;
+
+        if ($associatedServers->count() > 0) {
+            return true;
+        }
+
+        return false;
     }
 }
